Import path fixes for refactor
